### 2024-02-16 #159 @robelgeda
PetroFit
Submitting Author: Robel Geda (@robelgeda)
All current maintainers: (@robelgeda, @crawfordsm)
Package Name: PetroFit
One-Line Description of Package:The PetroFit Project is an open-source effort to develop end-to-end tools for making accurate photometric measurements, estimating morphological properties, and fitting 2D models to galaxy images.
Repository Link:  https://github.com/PetroFit/petrofit
Version submitted:  v0.5.0 (will be updated to v1.0.0 after review)
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14962700.svg)](https://doi.org/10.5281/zenodo.14962700)
JOSS DOI: N/A
Version accepted: [v0.6.0](https://github.com/PetroFit/petrofit/releases/tag/v0.6.0)
Date accepted: 02/28/2025

---

The goal of the `PetroFit` Python package, which is based on  [Astropy](https://www.astropy.org)  and [Photutils](https://photutils.readthedocs.io/en/stable/), is to provide specialized tools for the astronomical community. It is designed for calculating Petrosian properties, such as radii and concentration indices of galaxies, as well as fitting galaxy light profiles. In particular, `PetroFit` includes tools for performing accurate photometry, segmentations, Petrosian profiling, and Sérsic fitting.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.
