### 2024-02-20 #159 @lwasser
hey there @robelgeda 👋 i suspect this is an astropy package. can you kindly confirm? if so i'll work on updating the template with the astropy "check" this week as well. cc:  @isabelizimm  (we can chat in slack about this too! it will be our first astropy package) 🚀
