### 2024-02-24 #159 @isabelizimm
---
---

Welcome

1. Update `README.md` to include installation guide
2. In the installation, there is a section that states:

> Before installing PetroFit, you may need to install the required dependencies. You can do this using the requirements file located in the top directory of the repository. ...

package managers will install all the required dependencies of a package, as long as they are specified (in this package's case, in the `setup.cfg` in the `install_requires` section). If this is referring to installing tools for people to develop the PetroFit package, it should go in the `CONTRIBUTING.md` or Developer install instructions.

3. Add zenodo badge (I see one on the [Citing and Credits](https://petrofit.readthedocs.io/en/latest/citing.html)) page in docs to `README.md`
4. Add a docstring to functions in the public API:
	- [get_default_gen_sersic_bounds](https://github.com/PetroFit/petrofit/blob/db7b55159f493c1c11704f78475c01750f20da34/petrofit/modeling/models.py#L76)
	- [petrosian_profile](https://github.com/PetroFit/petrofit/blob/db7b55159f493c1c11704f78475c01750f20da34/petrofit/modeling/models.py#L583)

Suggestions/random things I noticed (non-blocking to review):

- In `ci_tests.yml`, the `prefix` parameter looks to be unused. Also, it looks like you have `matrix.allow_failure` always set to false, so you could probably just set `continue-on-error` to be false. I'm not sure if this CI suite is something you plan to customize/expand more and these are set in anticipation of later changes, but at the moment it's quite repetitive.

I have never seen an `Imposter syndrome disclaimer` before, and I have to say it made me a little emotional--I love it **so much** 🤍 what an awesome note to add to a README!
