### 2024-02-27 cb03a86
pyOpenSci: remove `update_default_config` call
