### 2024-02-27 bb1c288
pyOpenSci: add Zonodo tag to README.rst
