### 2024-02-27 a332f1e
pyOpenSci: cleanup docstrings in `models.py`
