### 2024-02-27 9851e86
pyOpenSci: imporve installation instruct
