### 2024-02-27 #159 @robelgeda
Hi @isabelizimm, thank you for taking the time and your kind words! I am also sorry about the late reply, I wanted to see if I can update things quickly.

I found your suggestions very useful and have updated the main branch. For the installation, I would need to make a release to upload to pip but would prefer to wait until the review process is complete. Other than that, we should be all set. Here are some quick notes and comments:

- I added a quick section about installing using pip.
- Related to this, the package should be able to install using pip alone now. Astronomers tend to be a fan of requirement files especially because they sometimes have cluttered environments. That being said, I am a fan of not needing one for installation as well!
- I have added the zenodo tag to our README.
- I cleaned up the functions in `models.py` and added the docstrings as requested.
- Thank you for pointing out the `ci_tests.yml` items, I cleaned up the file!
- I love that AstroPy has added the section about imposter syndrome, I believe it was first adopted by `MetPy`. I think its a challenge that often gets overlooked and one I am well acquainted with so I am proud to have it on our intro!

Thanks again!
