### 2024-06-04 #159 @robelgeda
Hi @dhomeier, I hope all is well. I wanted to check in to see if we have found reviewers for the package. There is updates that we wanted to do but I wanted to get them started after the astropy review process. What does the timeline look like for this review so we can plan accordingly. - Thank you for your time.
