+++ tox.ini
@@ -1,14 +1,14 @@
 [tox]
 envlist =
-    py{310}-test{-alldeps}
-    py{310}-test-numpy{122}
-    py{310}-test-astropy{53}
+    py{312}-test{-alldeps}
+    py{312}-test-numpy{122}
+    py{312}-test-astropy{61}
     build_docs
     linkcheck
 
 requires =
-    setuptools >= 61.0.0
-    pip >= 22.0.0
+    setuptools >= 72.2.0
+    pip >= 24.0.0
 
 isolated_build = true
 
@@ -39,14 +39,16 @@ description =
     run tests
     alldeps: with all optional dependencies
     numpy122: with numpy 1.22.*
-    astropy53: with astropy 5.3.*
+    numpy2: with numpy 2.0.*
+    astropy61: with astropy 6.1.*
     astropylts: with the latest astropy LTS
 
 # The following provides some specific pinnings for key packages
 deps =
     -r{toxinidir}/requirements.txt
+    numpy2: numpy==2.0.*
     numpy122: numpy==1.22.*
-    astropy53: astropy==5.3.*
+    astropy61: astropy==6.1.*
 
 # The following indicates which extras_require from setup.cfg will be installed
 extras =
