+++ setup.cfg
@@ -14,9 +14,10 @@ github_project = PetroFit/petrofit
 [options]
 zip_safe = False
 packages = find:
-python_requires = >=3.7
+python_requires = >=3.12
 setup_requires = setuptools_scm
 install_requires =
+    Cython
     numpy
     matplotlib
     astropy
@@ -25,7 +26,6 @@ install_requires =
     scikit-learn
     scikit-image
     photutils
-
 
 [options.extras_require]
 test =
