+++ CHANGES.md
@@ -11,7 +11,7 @@
 
 - Update rtd to use `ubuntu-22.04`. #200
 - `update_default_config` removed since `astropy` no longer needs it. #205
-- Update rtd to Use Python 3.11.8. #206
+- Update rtd to Use `Python 3.11.8`. #206
 - Update installation requiements for use with `python 3.12`. #209
 - Remove `astropy_helpers` from requirements. #209
 
