### 2024-08-16 6a449ae
update numpy version in tests
