+++ .github/workflows/ci_tests.yml
@@ -31,8 +31,8 @@ jobs:
 
           - os: ubuntu-latest
             python: '3.12'
-            tox_env: 'py312-test-numpy122'
-            prefix: 'py312-test-numpy122'
+            tox_env: 'py312-test-numpy2'
+            prefix: 'py312-test-numpy2'
 
           - os: ubuntu-latest
             python: '3.12'
@@ -51,7 +51,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install tox
+        python -m pip install tox setuptools
+        python -m pip install --upgrade setuptools
     - name: Test with Tox
       run: |
         export PY_COLORS=1
