### 2024-10-06 #159 @nden
- author email is missing

          The package has a CI badge but does not have one for test coverage.

           There is no repostatus badge
        It does not have a list of supported Python versions.
        This is something that tripped me initially. I installed it in a Python 3.11 environment without any warnings or errors and it failed to import. After realizing that all tests use Python 3.12 and installing it in a 3.12 env it works.

   Citation information is included in the documentation. I recommend adding it to the README file to make it easily discoverable.

   **Note**: The package uses the now archived astropy template structure. It is very close to the pyOpenSci packaging guidelines. I recommend updating and removing the astropy specific bits.

    The code formatting is standard and follows PEP 8 guidelines. However, running pylint generates a lot of messages pointing to issues which can be easily fixed (though they do not cause functional issues).

---

The package is in a very good shape and functional. Recommendations which will improve it are

- Add a list of supported Python versions to the README
- Add coverage test, report and badge
- Add a pylint Ci test
- Update the package structure to remove files from the old astropy template and modernize it according to the pyOpenSci packaging guidelines linked above
