### 2024-10-06 #159 @dhomeier
Thanks for your review @nden!
To chime in on the supported versions, the `python_requires = >=3.12` is reflected on PyPI, and a git install with 3.10 or 3.12 gives me 0.5.0 instead; this however is not compatible with astropy 6.x.
I do not see any obvious changes in the code in 0.5.1 that would break compatibility with 3.10 or 3.11, so is it really necessary to stop supporting these versions already? Ultimately this is of course your decision, but it's foreseeable that this will lead to errors and confusion for users who simply try to install it in an existing 3.10 or 3.11 environment. Generally it's recommendable to stay in sync with the dependency guidelines of [SPEC 0](https://scientific-python.org/specs/spec-0000/) that Numpy (and Astropy) follow as well.
