### 2024-10-06 #159 @dhomeier
@robelgeda we have also checked the conditions for a JOSS publication following up on your paper in the AJ, and found that it is still possible to submit a publication focussing on the code 2 years later. So if you wish to still add a `paper.md` within the next 2 weeks, we can probably get it reviewed without much additional delay. If you do reference the AJ publication, you should probably point out changes since the version described there.
