### 2024-10-21 #159 @kyleaoman
---
4

I've now finished my review. Overall the package seems to be in reasonably good shape. There are a few things that I think should be addressed before I sign off:
 - PetroFit/petrofit#213
 - PetroFit/petrofit#214
 - PetroFit/petrofit#215
 - PetroFit/petrofit#216
 - PetroFit/petrofit#218
 - PetroFit/petrofit#236

I'll also highlight an existing issue from the authors PetroFit/petrofit#104 : there are tests and they do cover the essential functions of the package at a level adequate to pass review, but the overall/long-term health of the package would probably benefit from better test coverage.

I've also kept some notes of other things that I've noticed, but for the purpose of this review these should be considered optional:
 - PetroFit/petrofit#212
 - PetroFit/petrofit#219
 - PetroFit/petrofit#220
