### 2024-10-21 PetroFit/petrofit#212 @kyleaoman
Add PR & issue templates
Github allows providing templates for issues and/or pull requests by providing templates under `project_root/.github/ISSUE_TEMPLATE` and similar. Consider making use of these features.

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is optional.
