### 2024-10-21 PetroFit/petrofit#213 @kyleaoman
readme badges
Some badges are missing from the readme:

- [x] A [repostatus.org](https://www.repostatus.org/) badge should be added to the readme file.
- [x] A badge listing supported python versions (automatically obtained from the repo metadata) should be added to the readme file.

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is required.
