### 2024-10-21 PetroFit/petrofit#214 @kyleaoman
code linting in continuous integration
Continuous integration should check that code conforms to a consistent style. Some suggestions:

- [ ] At a minimum check PEP8 compliance with a tool such as [flake8](https://flake8.pycqa.org/en/latest/).
- [ ] Recommend adopting a formatting style such as [black](https://black.readthedocs.io/en/stable/) and checking in CI.
- [ ] Consider also checking docstrings. They appear to loosely conform to the numpy style. This can be checked (and the level of compliance configured) with a tool such as [numpydoc](https://numpydoc.readthedocs.io/en/latest/validation.html). At a minimum checking for consistency between arguments in function definitions and in docstring is a good idea.

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is required.
