### 2024-10-21 PetroFit/petrofit#215 @kyleaoman
citation information should be included in readme
There is a `CITATION.rst` file present, but citation information is not mentioned/linked in the readme file.

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is required.
