### 2024-10-21 PetroFit/petrofit#216 @kyleaoman
include example(s) in readme
The PyOpenSci guidelines specify that the readme should include: "Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file."

This seems perhaps a bit too stringent as a requirement, but either directly including a simple example or linking to a "quickstart guide" type page in the documentation would be a good idea.

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is required.
