### 2024-10-21 PetroFit/petrofit#218 @kyleaoman
local docs build via tox fails
Attempting to run CI locally results in:
```
  py312-test-alldeps: OK (27.14=setup[2.58]+cmd[0.16,24.41] seconds)
  py312-test-numpy2: OK (26.16=setup[1.46]+cmd[0.16,24.54] seconds)
  py312-test-astropy61: OK (26.48=setup[1.42]+cmd[0.16,24.90] seconds)
  build_docs: FAIL code 2 (13.00=setup[1.40]+cmd[0.15,11.45] seconds)
  linkcheck: FAIL code 2 (13.22=setup[1.42]+cmd[0.16,11.64] seconds)
  evaluation failed :( (106.04 seconds)
```
A detailed log is copied below.

Steps to reproduce:
 - `. ~/.venv/petrofit/bin/activate` (a new venv with python 3.12.7).
 - cd to root directory (`commit fcf9c8a0c66cde47e79920ab7e5b888c26f9e263 (HEAD -> main, tag: v0.5.1, origin/main, origin/HEAD)`)
 - `pip install -r requirements.txt`
 - `pip install -e .`
 - `pip install --upgrade pip`
 - `tox`

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is required.
