### 2024-10-21 PetroFit/petrofit#218 @kyleaoman
```
.pkg: _optional_hooks> python /home/txwx36/.venv/petrofit/lib64/python3.12/site-packages/pyproject_api/_backend.py True setuptools.build_meta
.pkg: get_requires_for_build_sdist> python /home/txwx36/.venv/petrofit/lib64/python3.12/site-packages/pyproject_api/_backend.py True setuptools.build_meta
.pkg: get_requires_for_build_wheel> python /home/txwx36/.venv/petrofit/lib64/python3.12/site-packages/pyproject_api/_backend.py True setuptools.build_meta
.pkg: prepare_metadata_for_build_wheel> python /home/txwx36/.venv/petrofit/lib64/python3.12/site-packages/pyproject_api/_backend.py True setuptools.build_meta
.pkg: build_sdist> python /home/txwx36/.venv/petrofit/lib64/python3.12/site-packages/pyproject_api/_backend.py True setuptools.build_meta
py312-test-alldeps: install_package> python -I -m pip install --force-reinstall --no-deps /home/txwx36/code/petrofit/.tox/.tmp/package/22/petrofit-0.5.1.tar.gz
py312-test-alldeps: commands[0] /home/txwx36/code/petrofit/.tmp/py312-test-alldeps> pip freeze
alabaster==1.0.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.4
astropy-iers-data==0.2024.10.21.0.33.21
astropy-sphinx-theme==1.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.6.4
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docutils==0.21.2
executing==2.1.0
extension-helpers==1.2.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hypothesis==6.115.3
idna==3.10
imageio==2.36.0
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.5
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.2
numpydoc==1.8.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
petrofit @ file:///home/txwx36/code/petrofit/.tox/.tmp/package/22/petrofit-0.5.1.tar.gz#sha256=bfe1e253ecaf992f17cc1d7c0f36e7aa63abc70df66cad94a57d2ec727c4c0a2
pexpect==4.9.0
photutils==2.0.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyerfa==2.0.1.4
Pygments==2.18.0
pyparsing==3.2.0
pyproject-api==1.8.0
pytest==8.3.3
pytest-arraydiff==0.6.1
pytest-astropy==0.11.0
pytest-astropy-header==0.2.2
pytest-cov==5.0.0
pytest-doctestplus==1.2.1
pytest-filter-subpackage==0.2.0
pytest-mock==3.14.0
pytest-remotedata==0.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regions==0.10
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-astropy==1.9.1
sphinx-automodapi==0.18.0
sphinx-gallery==0.18.0
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
tornado==6.4.1
tox==4.23.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.27.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
py312-test-alldeps: commands[1] /home/txwx36/code/petrofit/.tmp/py312-test-alldeps> pytest --pyargs petrofit /home/txwx36/code/petrofit/docs
============================================================ test session starts ============================================================
cachedir: .tox/py312-test-alldeps/.pytest_cache
Running tests in petrofit docs.

Date: 2024-10-21T15:16:13

Platform: Linux-6.10.6-200.fc40.x86_64-x86_64-with-glibc2.39

Executable: /home/txwx36/code/petrofit/.tox/py312-test-alldeps/bin/python

Full Python Version:
3.12.7 (main, Oct  1 2024, 00:00:00) [GCC 14.2.1 20240912 (Red Hat 14.2.1-3)]

encodings: sys: utf-8, locale: UTF-8, filesystem: utf-8
byteorder: little
float info: dig: 15, mant_dig: 15

Package versions:
Numpy: 2.1.2
Scipy: 1.14.1
Matplotlib: 3.9.2
h5py: not available
Pandas: not available

Using Astropy options: remote_data: none.

rootdir: /home/txwx36/code/petrofit
configfile: setup.cfg
plugins: anyio-4.6.2.post1, doctestplus-1.2.1, astropy-header-0.2.2, hypothesis-6.115.3, remotedata-0.4.1, mock-3.14.0, filter-subpackage-0.2.0, arraydiff-0.6.1, cov-5.0.0, astropy-0.11.0
collected 28 items

../../.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_fitting.py ....                                        [ 14%]
../../.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py ....                                      [ 28%]
../../.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py ....                                     [ 42%]
../../.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py ........                               [ 71%]
../../.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_utils.py ........                                      [100%]

============================================================= warnings summary ==============================================================
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_fitting.py:234: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py::test_petrosian
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py:97: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py::test_petrosian
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py:100: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py::test_petrosian
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py:103: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py::test_corrections
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py:140: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py: 791 warnings
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py: 15 warnings
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py: 120 warnings
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/photometry.py:192: DeprecationWarning: Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future. Ensure you extract a single element from your array before performing this operation. (Deprecated NumPy 1.25.)
    aperture_area = float(np.round(aperture_area, 6))

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py: 791 warnings
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py: 15 warnings
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py: 120 warnings
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/photometry.py:193: DeprecationWarning: Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future. Ensure you extract a single element from your array before performing this operation. (Deprecated NumPy 1.25.)
    photometric_value = float(np.round(photometric_value, 6))

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py::test_plot_apertures
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py:19: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py::test_plot_apertures
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py:23: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py::test_plot_apertures
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py:27: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py::test_plot_apertures
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py:31: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py::test_radial_photometry
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_photometry.py:70: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_make_catalog_no_deblend
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_plot_segments
.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_plot_segments
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib64/python3.12/site-packages/matplotlib/cbook.py:733: DeprecationWarning: __array__ implementation doesn't accept a copy keyword, so passing copy=False failed. __array__ must implement 'dtype' and 'copy' keyword arguments.
    x = np.array(x, subok=True, copy=copy)

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_make_catalog_no_deblend
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py:74: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_plot_segments
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py:151: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_plot_segments
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py:154: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_plot_segments
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py:157: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py::test_source_photometry
  /home/txwx36/code/petrofit/.tox/py312-test-alldeps/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py:189: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
==================================================== 28 passed, 1872 warnings in 24.40s =====================================================
py312-test-alldeps: OK ✔ in 27.23 seconds
py312-test-numpy2: install_package> python -I -m pip install --force-reinstall --no-deps /home/txwx36/code/petrofit/.tox/.tmp/package/23/petrofit-0.5.1.tar.gz
py312-test-numpy2: commands[0] /home/txwx36/code/petrofit/.tmp/py312-test-numpy2> pip freeze
alabaster==1.0.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.4
astropy-iers-data==0.2024.10.21.0.33.21
astropy-sphinx-theme==1.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.6.4
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docutils==0.21.2
executing==2.1.0
extension-helpers==1.2.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hypothesis==6.115.3
idna==3.10
imageio==2.36.0
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.5
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.2
numpydoc==1.8.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
petrofit @ file:///home/txwx36/code/petrofit/.tox/.tmp/package/23/petrofit-0.5.1.tar.gz#sha256=bfe1e253ecaf992f17cc1d7c0f36e7aa63abc70df66cad94a57d2ec727c4c0a2
pexpect==4.9.0
photutils==2.0.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyerfa==2.0.1.4
Pygments==2.18.0
pyparsing==3.2.0
pyproject-api==1.8.0
pytest==8.3.3
pytest-arraydiff==0.6.1
pytest-astropy==0.11.0
pytest-astropy-header==0.2.2
pytest-cov==5.0.0
pytest-doctestplus==1.2.1
pytest-filter-subpackage==0.2.0
pytest-mock==3.14.0
pytest-remotedata==0.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regions==0.10
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-astropy==1.9.1
sphinx-automodapi==0.18.0
sphinx-gallery==0.18.0
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
tornado==6.4.1
tox==4.23.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.27.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
py312-test-numpy2: commands[1] /home/txwx36/code/petrofit/.tmp/py312-test-numpy2> pytest --pyargs petrofit /home/txwx36/code/petrofit/docs
============================================================ test session starts ============================================================
cachedir: .tox/py312-test-numpy2/.pytest_cache
Running tests in petrofit docs.

Date: 2024-10-21T15:16:40

Platform: Linux-6.10.6-200.fc40.x86_64-x86_64-with-glibc2.39

Executable: /home/txwx36/code/petrofit/.tox/py312-test-numpy2/bin/python

Full Python Version:
3.12.7 (main, Oct  1 2024, 00:00:00) [GCC 14.2.1 20240912 (Red Hat 14.2.1-3)]

encodings: sys: utf-8, locale: UTF-8, filesystem: utf-8
byteorder: little
float info: dig: 15, mant_dig: 15

Package versions:
Numpy: 2.0.2
Scipy: 1.14.1
Matplotlib: 3.9.2
h5py: not available
Pandas: not available

Using Astropy options: remote_data: none.

rootdir: /home/txwx36/code/petrofit
configfile: setup.cfg
plugins: anyio-4.6.2.post1, doctestplus-1.2.1, astropy-header-0.2.2, hypothesis-6.115.3, remotedata-0.4.1, mock-3.14.0, filter-subpackage-0.2.0, arraydiff-0.6.1, cov-5.0.0, astropy-0.11.0
collected 28 items

../../.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_fitting.py ....                                         [ 14%]
../../.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py ....                                       [ 28%]
../../.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_photometry.py ....                                      [ 42%]
../../.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_segmentation.py ........                                [ 71%]
../../.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_utils.py ........                                       [100%]

============================================================= warnings summary ==============================================================
.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_fitting.py::test_psf_sampling
  /home/txwx36/code/petrofit/.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_fitting.py:234: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py::test_petrosian
  /home/txwx36/code/petrofit/.tox/py312-test-numpy2/lib/python3.12/site-packages/petrofit/tests/test_petrosian.py:97: UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown
    plt.show()

[…]
