### 2024-10-21 ae348f2
Missing second backslash in docstring math.
