+++ petrofit/segmentation.py
@@ -98,7 +98,7 @@ def get_amplitude_at_r(r, image, x0, y0, ellip, theta):
         The position angle (in radians) of the semimajor axis in
         relation to the positive x axis of the image array (rotating
         towards the positive y axis). Position angles are defined in the
-        range :math:`0 < PA <= \pi`. Avoid using as starting position
+        range :math:`0 < PA <= \\pi`. Avoid using as starting position
         angle of 0., since the fit algorithm may not work properly.
         When the ellipses are such that position angles are near either
         extreme of the range, noise can make the solution jump back and
