### 2024-10-21 PetroFit/petrofit#219 @kyleaoman
bullet list rendering glitch in docs
In this docs page: https://petrofit.readthedocs.io/en/stable/photometry.html

There is a bullet list near "four plots are displayed" that does not seem to have rendered correctly (though it looks fine if I open the notebook in the repository).

This issue is part of my PyOpenSci review. For review pass/fail purposes, resolving it is optional.
