### 2024-10-22 238a5a0
Small corrections/typo fixes in docstrings.
