### 2024-10-25 #159 @dhomeier
Looking forward to your updates @robelgeda!
Note on the citation information: Astropy has already recommended using a `CITATION` file for some time; we are currently discussing how to best update the pyOpenSci packaging guide to reflect this. A useful addition would certainly be a link in the README to the former file.
