+++ pyproject.toml
@@ -1,7 +1,251 @@
+# [build-system]
+
+# requires = ["setuptools>=72",
+#             "setuptools_scm",
+#             "wheel"]
+
+# build-backend = 'setuptools.build_meta'
+
+
+################################################################################
+# Build Configuration
+################################################################################
+
 [build-system]
+build-backend = "hatchling.build"
+requires = ["hatchling", "hatch-vcs"]
 
-requires = ["setuptools>=72",
-            "setuptools_scm",
-            "wheel"]
+################################################################################
+# Project Configuration
+################################################################################
 
-build-backend = 'setuptools.build_meta'
+[project]
+name = "petrofit"
+
+dynamic = ["version"]
+
+description = "Python package for calculating Petrosian properties and fitting galaxy light profiles."
+authors = [
+    { name = "The PetroFit Team", email = "" },
+]
+license = "BSD-3-Clause"
+readme = {"file" = "README.md", "content-type" = "text/markdown"}
+requires-python = ">=3.12"
+# Please consult https://pypi.org/classifiers/ for a full list.
+classifiers = [
+    "Development Status :: 2 - Pre-Alpha",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: BSD License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Scientific/Engineering",
+    "Typing :: Typed",
+]
+
+keywords = ["fitting", "galaxy", "light", "profile", "petrosian", "photometry"]
+dependencies = [
+    "Cython",
+    "numpy",
+    "matplotlib",
+    "astropy",
+    "scipy",
+    "pyyaml",
+    "scikit-learn",
+    "scikit-image",
+    "photutils",
+]
+
+[project.urls]
+Homepage = "https://github.com/PetroFit/petrofit"
+"Source Code" = "https://github.com/PetroFit/petrofit"
+"Bug Tracker" = "https://github.com/PetroFit/petrofit/issues"
+Documentation = "https://petrofit.readthedocs.io"
+Download = "https://pypi.org/project/petrofit/#files"
+
+[project.optional-dependencies]
+dev = [
+    "hatch",
+]
+
+################################################################################
+# Tool Configuration
+################################################################################
+
+[tool.hatch.build]
+only-packages = true
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/petrofit"]
+[tool.hatch.build.hooks.vcs]
+version-file = "src/petrofit/_version.py"
+
+[tool.hatch.version]
+source = "vcs"
+[tool.pytest.ini_options]
+testpaths = ["petrofit"]
+markers = ["raises"]
+astropy_header = true
+doctest_plus = "enabled"
+text_file_format = "rst"
+addopts = "--doctest-rst"
+
+[tool.coverage.paths]
+source = [
+    "src/petrofit",
+    "*/site-packages/petrofit",
+]
+
+[tool.coverage.run]
+branch = true
+parallel = true
+omit = [
+    "src/petrofit/_version.py",
+]
+
+[tool.coverage.report]
+exclude_lines = ["pragma: no cover"]
+precision = 2
+[tool.ruff]
+line-length = 88
+
+[tool.ruff.lint]
+select = ["ALL"]
+ignore = [
+    "D107",  # 'Missing docstring in __init__' ignored because pydoclint wants us to document the class instead.
[…]
