+++ .github/workflows/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,4 @@
+- [ ] fix #(issue number)
+- [ ] description of feature/fix
+- [ ] tests added/passed
+- [ ] add an entry to the [changelog](../CHANGELOG.md)
