+++ .github/workflows/workflows/cron.yml
@@ -1,0 +1,10 @@
+name: Cron Test
+
+on:
+  schedule:
+  # Run every Tuesday at 10:30.
+  - cron: '30 10 * * 2'
+
+jobs:
+  prerequisites:
+    uses: ./.github/workflows/test.yml
