+++ .github/workflows/workflows/main.yml
@@ -1,0 +1,13 @@
+name: CI
+
+on:
+  push:
+    branches:
+    - main
+  pull_request:
+    branches:
+    - main
+
+jobs:
+  prerequisites:
+    uses: ./.github/workflows/test.yml
