+++ .github/workflows/ISSUE_TEMPLATE/01-bug-report.yml
@@ -1,0 +1,60 @@
+name: Bug report
+description: Report a problem to help improve this project
+title: "[BUG] "
+labels: [bug, triage]
+body:
+  - type: textarea
+    attributes:
+      label: Problem description
+      description: |
+        A concise description of what you're experiencing.
+
+        Please explain:
+
+          * **what** you tried to achieve,
+          * **how** you went about it (referring to the code sample), and
+          * **why** the current behaviour is a problem and what output you expected instead.
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Code sample
+      description: >
+        Create a [minimal, complete, verifiable example](https://stackoverflow.com/help/mcve).
+        Please, paste your code between the ``` tickmarks below or link to a [gist](https://gist.github.com/).
+      value: |
+        Code run:
+
+        ```python
+        ```
+
+        Traceback:
+
+        ```text
+        ```
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Environment
+      description: >
+        Please paste the output of running `depinfo --markdown petrofit`
+        in your environment between the `details` tags below.
+      value: |
+        <details>
+
+        </details>
+    validations:
+      required: true
+
+  - type: textarea
+    attributes:
+      label: Anything else?
+      description: |
+        Links? References? Anything that will give us more context about the issue you are encountering!
+
+        Tip: You can attach images or log files by clicking this area to highlight it and then dragging files in.
+    validations:
+      required: false
