### 2024-10-30 92b9e5e
remove PR template, we want 0 friction
