+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,4 +1,0 @@
-- [ ] fix #(issue number)
-- [ ] description of feature/fix
-- [ ] tests added/passed
-- [ ] add an entry to the [changelog](../CHANGELOG.md)
