+++ .github/ISSUE_TEMPLATE/03-feature-request.yml
@@ -5,7 +5,7 @@ labels: [enhancement]
 body:
   - type: textarea
     attributes:
-      label: Problem
+      label: Feature or Problem
       description: >
         A clear and concise description of what you are trying to achieve.
       placeholder: >
