### 2024-10-30 fe5eb28
fix install issues and remove docker
