+++ pyproject.toml
@@ -42,7 +42,6 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering",
-    "Typing :: Typed",
 ]
 
 keywords = ["fitting", "galaxy", "light", "profile", "petrosian", "photometry"]
@@ -127,10 +126,16 @@ ignore = [
     "D203",  # '1 blank line required before class docstring' ignored because we want no blank line.
     "D212",  # 'Multi-line docstring summary should start at the first line' ignored because we want the summary to start on the second line.
     "D407",  # 'Missing dashed underline after section' ignored because Google style docstrings don't underline.
+    "ANN001",  # Ignore missing type annotations for arguments
+    "ANN101",  # Ignore missing type annotation for `self` in methods
+    "ANN201",  # Ignore missing type annotation for return values
     "ANN002",  # 'Missing type annotation for {*args} in method'.
     "ANN003",  # 'Missing type annotation for {*kwargs} in method'.
     "ANN101",  # 'Missing type annotation for {self} in method'.
     "ANN102",  # 'Missing type annotation for {cls} in classmethod'.
+    "Q000", # 'Remove bad quotes' ignored because we want to use double quotes.
+    "FBT002",  # Boolean default positional argument in function definition"
+
 ]
 exclude = [
     "petrofit/_version.py",
