+++ docs/conf.py
@@ -13,11 +13,9 @@ except ImportError:
     sys.exit(1)
 
 # Get configuration information from setup.cfg
-from configparser import ConfigParser
-conf = ConfigParser()
-
-conf.read([os.path.join(os.path.dirname(__file__), '..', 'setup.cfg')])
-setup_cfg = dict(conf.items('metadata'))
+import tomllib
+with open("pyproject.toml", "rb") as f:
+    config = tomllib.load(f)
 
 # -- General configuration ----------------------------------------------------
 
@@ -36,11 +34,12 @@ nbsphinx_execute = 'auto'
 # -- Project information ------------------------------------------------------
 
 project = "PetroFit"
-author = setup_cfg['author']
+author = config["project"]['authors'][0]['name']
+module_name = config["project"]['name']
 copyright = '{0}, {1}'.format(
-    datetime.datetime.now().year, setup_cfg['author'])
-import_module(setup_cfg['name'])
-package = sys.modules[setup_cfg['name']]
+    datetime.datetime.now().year, author)
+import_module(module_name)
+package = sys.modules[module_name]
 release = package.__version__
 version = release
 if "dev" in release:
@@ -69,15 +68,6 @@ latex_documents = [('index', project + '.tex', project + u' Documentation',
 
 man_pages = [('index', project.lower(), project + u' Documentation',
               [author], 1)]
-
-# -- Options for the edit_on_github extension ---------------------------------
-
-if setup_cfg.get('edit_on_github').lower() == 'true':
-    extensions += ['sphinx_astropy.ext.edit_on_github']
-    edit_on_github_project = setup_cfg['github_project']
-    edit_on_github_branch = "main"
-    edit_on_github_source_root = ""
-    edit_on_github_doc_root = "docs"
 
 # -- Resolving issue number to links in changelog -----------------------------
 
