+++ .readthedocs.yaml
@@ -5,21 +5,17 @@
 # Required
 version: 2
 
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "mambaforge-4.10"
-
-conda:
-  environment: environment.yml
-
 sphinx:
   builder: html
   configuration: docs/conf.py  # This should point to your conf.py
 
-python:
-  install:
-    - method: pip
-      path: .
-      extra_requirements:
-        - docs
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.12"
+  commands:
+    - sudo apt-get update
+    - sudo apt-get install -y pandoc  
+    - pip install --upgrade pip setuptools wheel
+    - pip install hatch
+    - hatch run docs:build
