+++ .readthedocs.yaml
@@ -13,9 +13,9 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.12"
+  apt_packages:
+    - pandoc
   commands:
-    - sudo apt-get update
-    - sudo apt-get install -y pandoc  
     - pip install --upgrade pip setuptools wheel
     - pip install hatch
     - hatch run docs:build
