### 2024-10-30 8fc45e4
remove astropy helper
