+++ petrofit/__init__.py
@@ -1,11 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-
-# Packages may add whatever they like to this file, but
-# should keep this content at the top.
-# ----------------------------------------------------------------------------
-from ._astropy_init import *   # noqa
-# ----------------------------------------------------------------------------
-
 from .petrosian import *
 from .photometry import *
 from .segmentation import *
