+++ .github/workflows/main.yml
@@ -3,10 +3,10 @@ name: CI
 on:
   push:
     branches:
-    - main
+    - '*' 
   pull_request:
     branches:
-    - main
+    - '*' 
 
 jobs:
   prerequisites:
