### 2024-10-30 ac2343f
make python badge dynamic
