### 2024-10-30 ecfe310
add citation info in README
