### 2024-10-30 fd65d8d
add example in README
