+++ petrofit/__init__.py
@@ -2,8 +2,7 @@
 try:
     from ._version import __version__
 except ImportError:
-    __version__ = "unknown"
-
+    __version__ = "N.A"
 
 from .petrosian import *
 from .photometry import *
