+++ petrofit/tests/test_segmentation.py
@@ -15,7 +15,7 @@ from matplotlib import pyplot as plt
 
 @pytest.fixture
 def sersic_2d_image(data_dir):
-    """fixture for Sersic 2D image """
+    """fixture for Sersic 2D image"""
     path = "sersic_2d_image.fits.gz"
     sersic_2d_path = os.path.join(data_dir, path)
     return fits.getdata(sersic_2d_path)
@@ -24,16 +24,19 @@ def sersic_2d_image(data_dir):
 @pytest.fixture
 def segm_and_cat(sersic_2d_image):
     """fixture for segmentation and catalog"""
-    image_mean, image_median, image_stddev = sigma_clipped_stats(sersic_2d_image, sigma=3)
+    image_mean, image_median, image_stddev = sigma_clipped_stats(
+        sersic_2d_image, sigma=3
+    )
     threshold = image_stddev * 3
-    npixels = 4 ** 2
+    npixels = 4**2
     return pf.make_catalog(
         sersic_2d_image,
         threshold=threshold,
         deblend=True,
         npixels=npixels,
         contrast=0.00,
-        plot=False)
+        plot=False,
+    )
 
 
 def test_make_catalog(segm_and_cat, sersic_2d_image):
@@ -57,20 +60,25 @@ def test_make_catalog(segm_and_cat, sersic_2d_image):
     assert segm_deblend.shape == sersic_2d_image.shape
 
     # Validate segmentation versus deblended segmentation
-    assert not np.array_equal(segm.data, segm_deblend.data)  # They should be different if deblending is effective
+    assert not np.array_equal(
+        segm.data, segm_deblend.data
+    )  # They should be different if deblending is effective
 
 
 def test_make_catalog_no_deblend(sersic_2d_image):
-    image_mean, image_median, image_stddev = sigma_clipped_stats(sersic_2d_image, sigma=3)
+    image_mean, image_median, image_stddev = sigma_clipped_stats(
+        sersic_2d_image, sigma=3
+    )
     threshold = image_stddev * 3
-    npixels = 4 ** 2
+    npixels = 4**2
     cat, segm, segm_deblend = pf.make_catalog(
         sersic_2d_image,
         threshold=threshold,
         deblend=False,
         npixels=npixels,
         contrast=0.00,
-        plot=True)
+        plot=True,
+    )
     plt.show()
     assert len(cat) == 1
     assert segm_deblend is None
@@ -80,9 +88,11 @@ def test_make_catalog_no_deblend(sersic_2d_image):
 def test_segmentation_functions(sersic_2d_image):
     """This tests the make_segments and deblend_segments functions"""
 
-    image_mean, image_median, image_stddev = sigma_clipped_stats(sersic_2d_image, sigma=3)
+    image_mean, image_median, image_stddev = sigma_clipped_stats(
+        sersic_2d_image, sigma=3
+    )
     threshold = image_stddev * 3
-    npixels = 4 ** 2
+    npixels = 4**2
 
     # Testing make_segments
     segm = pf.make_segments(sersic_2d_image, npixels=npixels, threshold=threshold)
@@ -93,7 +103,9 @@ def test_segmentation_functions(sersic_2d_image):
     assert len(np.unique(segm.data)) == 2  # account for background being labeled as 0
 
     # Testing deblend_segments
-    segm_deblend = pf.deblend_segments(sersic_2d_image, segm, npixels=npixels, contrast=0.00)
+    segm_deblend = pf.deblend_segments(
+        sersic_2d_image, segm, npixels=npixels, contrast=0.00
+    )
 
     assert isinstance(segm_deblend, SegmentationImage)
     assert segm_deblend.shape == sersic_2d_image.shape
[…]
