+++ petrofit/tests/test_fitting.py
@@ -24,11 +24,11 @@ def test_psf_convolved_image_model():
         ellip=0,
         theta=0,
         bounds={
-            'amplitude': (0., None),
-            'r_eff': (0, None),
-            'n': (0, 10),
-            'ellip': (0, 1),
-            'theta': (-2 * np.pi, 2 * np.pi),
+            "amplitude": (0.0, None),
+            "r_eff": (0, None),
+            "n": (0, 10),
+            "ellip": (0, 1),
+            "theta": (-2 * np.pi, 2 * np.pi),
         },
     )
 
@@ -45,7 +45,7 @@ def test_psf_convolved_image_model():
 
     # Make a PSFConvolvedModel2D
     psf_sersic_model = pf.PSFConvolvedModel2D(sersic_model, psf=PSF, oversample=None)
-    psf_sersic_model.fixed['psf_pa'] = True
+    psf_sersic_model.fixed["psf_pa"] = True
 
     # Make a PSFConvolvedModel2D image
     psf_sersic_model_image = pf.model_to_image(psf_sersic_model, imsize)
@@ -64,7 +64,8 @@ def test_psf_convolved_image_model():
 
     # Fit
     fitted_model, fit_info = pf.fit_model(
-        psf_sersic_image, psf_sersic_model,
+        psf_sersic_image,
+        psf_sersic_model,
         maxiter=10000,
         epsilon=1.4901161193847656e-10,
         acc=1e-9,
@@ -81,12 +82,8 @@ def test_psf_convolved_image_model():
 def test_make_grid():
     # Test default arguments
     x, y = pf.make_grid(3)
-    expected_x = np.array([[0., 1., 2.],
-                           [0., 1., 2.],
-                           [0., 1., 2.]])
-    expected_y = np.array([[0., 0., 0.],
-                           [1., 1., 1.],
-                           [2., 2., 2.]])
+    expected_x = np.array([[0.0, 1.0, 2.0], [0.0, 1.0, 2.0], [0.0, 1.0, 2.0]])
+    expected_y = np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0], [2.0, 2.0, 2.0]])
     assert np.allclose(x, expected_x)
     assert np.allclose(y, expected_y)
 
@@ -97,18 +94,23 @@ def test_make_grid():
 
     x, y = pf.make_grid(2, origin=(0.25, 0.25), factor=2)
     expected_x, expected_y = [
-        np.array([
-            [0., 0.5, 1., 1.5],
-            [0., 0.5, 1., 1.5],
-            [0., 0.5, 1., 1.5],
-            [0., 0.5, 1., 1.5]
-        ]),
-         np.array([
-             [0., 0., 0., 0.],
-             [0.5, 0.5, 0.5, 0.5],
-             [1., 1., 1., 1.],
-             [1.5, 1.5, 1.5, 1.5]
-         ])]
+        np.array(
+            [
+                [0.0, 0.5, 1.0, 1.5],
+                [0.0, 0.5, 1.0, 1.5],
+                [0.0, 0.5, 1.0, 1.5],
+                [0.0, 0.5, 1.0, 1.5],
+            ]
+        ),
+        np.array(
+            [
+                [0.0, 0.0, 0.0, 0.0],
+                [0.5, 0.5, 0.5, 0.5],
+                [1.0, 1.0, 1.0, 1.0],
+                [1.5, 1.5, 1.5, 1.5],
+            ]
+        ),
+    ]
     assert np.allclose(x, expected_x)
     assert np.allclose(y, expected_y)
 
[…]

[…]
