+++ petrofit/utils.py
@@ -11,19 +11,28 @@ from astropy.coordinates import SkyCoord
 from matplotlib import pyplot as plt
 
 
-
 __all__ = [
-    'match_catalogs', 'angular_to_pixel', 'pixel_to_angular',
-    'elliptical_area_to_r', 'circle_area_to_r', 'get_interpolated_values',
-    'closest_value_index', 'plot_target', 'cutout_subtract',
-    'hst_flux_to_abmag', 'make_radius_list', 'natural_sort', 'mpl_tick_frame',
-    'ellip_to_elong', 'elong_to_ellip'
+    "match_catalogs",
+    "angular_to_pixel",
+    "pixel_to_angular",
+    "elliptical_area_to_r",
+    "circle_area_to_r",
+    "get_interpolated_values",
+    "closest_value_index",
+    "plot_target",
+    "cutout_subtract",
+    "hst_flux_to_abmag",
+    "make_radius_list",
+    "natural_sort",
+    "mpl_tick_frame",
+    "ellip_to_elong",
+    "elong_to_ellip",
 ]
 
 
 def natural_sort(l):
     convert = lambda text: int(text) if text.isdigit() else text.lower()
-    alphanum_key = lambda key: [ convert(c) for c in re.split('([0-9]+)', key) ]
+    alphanum_key = lambda key: [convert(c) for c in re.split("([0-9]+)", key)]
     return sorted(l, key=alphanum_key)
 
 
@@ -38,7 +47,9 @@ def elong_to_ellip(elong):
 def make_radius_list(max_pix, n, log=False):
     """Make an array of radii of size n up to max_pix"""
     if log:
-        return np.logspace(0, np.log10(max_pix), num=n, endpoint=True, base=10.0, dtype=float, axis=0)
+        return np.logspace(
+            0, np.log10(max_pix), num=n, endpoint=True, base=10.0, dtype=float, axis=0
+        )
     else:
         return np.array([x * max_pix / n for x in range(1, n + 1)])
 
@@ -51,17 +62,17 @@ def hst_flux_to_abmag(flux, header):
     elif flux <= 0:
         return np.nan
 
-    PHOTFLAM = header['PHOTFLAM']
-    PHOTZPT = header['PHOTZPT']
-    PHOTPLAM = header['PHOTPLAM']
+    PHOTFLAM = header["PHOTFLAM"]
+    PHOTZPT = header["PHOTZPT"]
+    PHOTPLAM = header["PHOTPLAM"]
 
     STMAG_ZPT = (-2.5 * np.log10(PHOTFLAM)) + PHOTZPT
-    ABMAG_ZPT = STMAG_ZPT - (5. * np.log10(PHOTPLAM)) + 18.692
+    ABMAG_ZPT = STMAG_ZPT - (5.0 * np.log10(PHOTPLAM)) + 18.692
 
     return -2.5 * np.log10(flux) + ABMAG_ZPT
 
 
-def match_catalogs(ra_1, dec_1, ra_2, dec_2, unit='deg'):
+def match_catalogs(ra_1, dec_1, ra_2, dec_2, unit="deg"):
     """Wrapper for `SkyCoord.match_to_catalog_sky`"""
     cat1_coords = SkyCoord(ra=ra_1, dec=dec_1, unit=unit)
     cat2_coords = SkyCoord(ra=ra_2, dec=dec_2, unit=unit)
@@ -84,7 +95,7 @@ def pixel_to_angular(pixel_size, wcs):
     assert np.allclose(*pixel_scales)
     pixel_scale = pixel_scales[0] * wcs.wcs.cunit[0] / u.pix
 
-    if not hasattr(pixel_size, 'unit'):
[…]

[…]
