### 2024-10-31 750aaa1
fix cov install
