### 2024-10-31 1df5ac6
update CITATION.rst
