### 2024-10-31 ad521b9
do not fail docs on warn and add j lab
