+++ pyproject.toml
@@ -1,12 +1,3 @@
-# [build-system]
-
-# requires = ["setuptools>=72",
-#             "setuptools_scm",
-#             "wheel"]
-
-# build-backend = 'setuptools.build_meta'
-
-
 ################################################################################
 # Build Configuration
 ################################################################################
@@ -31,7 +22,6 @@ authors = [
 license = "BSD-3-Clause"
 readme = {"file" = "README.rst", "content-type" = "text/x-rst"}
 requires-python = ">=3.12"
-# Please consult https://pypi.org/classifiers/ for a full list.
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Science/Research",
@@ -42,11 +32,11 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering",
+    # Please consult https://pypi.org/classifiers/ for a full list.
 ]
 
 keywords = ["fitting", "galaxy", "light", "profile", "petrosian", "photometry"]
 dependencies = [
-    # Core Dependencies
     "pyyaml",
     'Cython',
     'numpy>=1.13',
@@ -82,6 +72,9 @@ dev = [
 [tool.setuptools_scm]
 write_to = "petrofit/_version.py"
 
+# hatch
+# -----
+
 [tool.hatch.build]
 only-packages = true
 
@@ -92,12 +85,19 @@ version-file = "petrofit/_version.py"
 
 [tool.hatch.version]
 source = "vcs"
+
+# pytest
+# ------
+
 [tool.pytest.ini_options]
 testpaths = ["petrofit"]
 markers = ["raises"]
 astropy_header = true
 doctest_plus = "enabled"
 text_file_format = "rst"
+
+# coverage
+# --------
 
 [tool.coverage.paths]
 source = [
@@ -115,6 +115,10 @@ omit = [
 [tool.coverage.report]
 exclude_lines = ["pragma: no cover"]
 precision = 2
+
+# ruff
+# ----
+
 [tool.ruff]
 line-length = 88
 
@@ -159,12 +163,19 @@ case-sensitive = true
 known-first-party = ["src", "petrofit"]
 lines-after-imports = 1
 
+# pydoclint
+# ---------
+
 [tool.pydoclint]
 style = "numpy"
 arg-type-hints-in-docstring = false
 check-return-types = false
 check-yield-types = false
 exclude = "_version.py"
+
+# other
+# -----
+
 # TODO: Adjust mypy configuration.
 #[tool.mypy]
 #plugins = [
@@ -182,9 +193,15 @@ exclude = "_version.py"
 #init_forbid_extra = true
 #init_typed = true
 #warn_required_dynamic_aliases = true
+
+
 ################################################################################
 # Hatch Environments
 ################################################################################
+
+# style
+# -----
+
 [tool.hatch.envs.style]
 description = """Check the style of the codebase."""
 dependencies = [
@@ -198,6 +215,9 @@ docstrings = "pydoclint"
 code = "ruff check {args}"
 format = "ruff format {args}"
 check = ["docstrings", "code"]
+
+# audit
+# -----
 
 [tool.hatch.envs.audit]
 description = """Check dependencies for security vulnerabilities."""
@@ -213,8 +233,14 @@ dependencies = [
     "mypy",
 ]
 
+# scripts
+# -------
+
 [tool.hatch.envs.types.scripts]
 check = "mypy petrofit"
+
+# docs
+# ----
 
 [tool.hatch.envs.docs]
[…]
