+++ .github/SUPPORT.md
@@ -1,4 +1,3 @@
 # Support
 
 * Start a [discussion topic](https://github.com/PetroFit/petrofit/discussions)
-* Email the authors or maintainers
