### 2024-11-01 07b53d4
fix link in README.rst
