### 2024-11-01 1f13a60
Update citation info in README.rst
