### 2024-11-01 7291a05
Move ads order in CITATION.rst
