### 2024-11-01 f78f5b1
Update link in README.rst
