### 2024-11-01 PetroFit/petrofit#214 @robelgeda
descoping due to docstrings
