### 2024-11-01 PetroFit/petrofit#214 @robelgeda
I have added `ruff` checks to the code but need time and careful attention to fix the other issues. It would either take more time or too quick to be fair to the end user to correct the doc and doc-string issues now.
