### 2024-11-01 #159 @robelgeda
@dhomeier the review was very useful for the project! I have added all the changes recommended except the doc updates. I plan to spend more time on docs once we have refactored the code to contribute upstream. That being said I have run `black` to comply with some of the PEP8 recommendations suggested in this review project.

For the JOSS submission, I think I will come back to it after this review since there are plans to introduce JAX based fitting code in the very near future (at least a branch working toward this goal). For the purposes of this review, I think its okay for us to process this application without the journal submission.
