### 2024-11-01 949bdd5
fix link in README.rst
