### 2024-11-01 #159 @robelgeda
I have also added a `Generative AI Policy` section to our README. This might not be common practice at the moment but it is important to supervise code added to our repo given that its main goals are science use cases. Generative AI can be used for docs, GitHub workflow (tickets, PRs...) and reports freely since these things are used to communicate and document changes to users as opposed to affecting the core function of the codebase.
