### 2024-11-02 b77381b
Minor fix in README.rst
