### 2024-11-26 #159 @dhomeier
Thanks @robelgeda for the extensive updates! @kyleaoman, @nden, it looks like all issues with the submission, especially those itemised under _Review comments_ and in `Petrofit`'s issues `#213-220` have been resolved with the exception of  some linting aspects, in particular involving docstrings.
Could you re-review the package to confirm that all mandatory criteria are now met? Thanks!
