### 2024-11-27 #159 @kyleaoman
Hi @dhomeier , @robelgeda . I had a look at the changes. Mostly looks good, but the updated test suite still fails, I've opened a new issue (added a link in my report).

Also not quite sure what to do with the docs item - if it's part of the pass/fail requirements for review then it should be done before passing, not promised for later. Some of my comments/suggestions for the docs are outside of the pass/fail scope, and those it's fine to defer of course, just trying to work out where to draw the line. I flagged a few functions with missing docstrings, and the tickbox on the review checklist says "function documentation for all functions user-facing functions". I suggest those get filled in to pass review from my side, if @dhomeier agrees.

Other than that everything looks pretty good to me!
