### 2024-12-02 PetroFit/petrofit#236 @dhomeier
I agree this does look like a code, in the first instance `psf_p` not being correctly initialised with an iterable (as default).
I also note that you were running against astropy 7.0.0, while the last tests on main were presumably run pre-7.0 – the only change that comes to mind here was astropy/astropy#16829 though...
