### 2024-12-02 #159 @dhomeier
Thanks for the quick reply @kyleaoman! I had missed that you had also flagged functions entirely without docstrings, and agree this is more problematic than the docs not yet conforming to a consistent numpydoc or other style.
The [pyOS Packaging Guide](https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/document-your-code-api-docstrings.html#python-package-api-documentation) asks for all of the user-facing API to be documented _within reason_, so there is certainly room for tolerance, but any function important for the normal workflow of using the package should have a docstring.
The case of the failing test with `PSFConvolvedModel2D.evaluate`
https://github.com/PetroFit/petrofit/blame/af30a253fb1ec45fd236e46d46230f81f0f21c8e/petrofit/modeling/models.py#L353-L356
seemingly not called with valid `psf_p` type may in fact emphasise this, the missing docs making it harder to see how `*params` should be set correctly.
