### 2025-02-07 c388163
patch PSFConvolvedModel2D
