### 2025-02-07 #159 @robelgeda
@kyleaoman, I am sorry about the late reply, I didn't realize your comments on the docstrings were about missing docstrings (I thought it was about formatting). I have run `pydocstyle --select=D101,D102,D103` and fixed all the missing docstrings, thank you for point that out.

The bug that was reported is part of the new astropy update, they updated the fitting module shortly after my last commit. I have added a minor fix and hatch test works. The CI caught the bug before I applied the fix as well.

@nden @dhomeier
