### 2025-02-10 #159 @dhomeier
Thanks, looking good to me too now!
@nden I've taken the liberty of checking the remaining open boxes for citation and badges, as those have been added (as well as coverage tests themselves). Do you have any final remarks on the pylint status – otherwise I'd consider those non-blocking.
