### 2025-02-28 #159 @dhomeier
Does not look like there are any objections, so I am moving this forward now.

---

🎉 [PetroFit](https://github.com/PetroFit/petrofit) has been approved by pyOpenSci! Thank you @robelgeda for submitting PetroFit and many thanks to @kyleaoman and @nden for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the [README.rst](https://github.com/PetroFit/petrofit/blob/main/README.rst) of [PetroFit](https://github.com/PetroFit/petrofit). The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
