### 2025-03-03 #159 @robelgeda
Thank you @dhomeier @kyleaoman @nden !!! The package looks and operates smoothly now!

@dhomeier I released v0.6.0 and all the items should be checked out now (I also took the survey)! Please let me know if I should do anything next on the Astropy end of things.
[Zonodo](https://zenodo.org/records/14962700)
[PyPi](https://pypi.org/project/petrofit/)
[Version 0.6.0 Note](https://github.com/PetroFit/petrofit/releases/tag/v0.6.0)

As mentioned above, we plan to release v1.0.0 after conversation with upstream Astropy maintainers about which part of the code can be generalized enough to live in their respective packages.
