### 2025-03-03 #159 @dhomeier
Thanks @robelgeda, I think this concludes all work on the package here, just leaving the post-review survey and, if you wish, sharing your experiences in a blog post.
Also if you, @crawfordsm, @nden or @kyleaoman wish to join the pyOpenSci Slack space, you are welcome to send us an email address to be signed up with.

I noticed one "trim trailing whitespace" error in the "Annotation" task of the [merge run](https://github.com/pyOpenSci/pyopensci.github.io/actions/runs/13596533379) (I think) for the PetroFit entry, will see if anything remains to be done to complete the https://www.pyopensci.org/python-packages.html listing.
