### 2025-05-02 #159 @hamogu
Petrofit is fully listed (and has been for a few weeks) so let's close this issue. Again, thanks to @robelgeda for submitting and @nden and @kyleaoman for reviewing!
